requirements.txt
# Core FastAPI dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Speech Services
elevenlabs>=0.2.24
deepgram-sdk>=3.0.1
openai-whisper>=20231117

# Twilio
twilio>=8.12.0

# WebSocket support
websockets>=12.0

# LLM and AI
openai>=1.12.0

# Utilities
requests>=2.31.0
chromadb>=0.5.0
psycopg2-binary>=2.9.9
pytest>=8.3.0
redis>=5.0.3

# ngrok
ngrok>=0.1.0

# ClaudeSync
claudesync>=0.1.0